+++ Editor/WakaTime/client/wakatime/stats.py
@@ -28,56 +28,19 @@ from pygments.util import ClassNotFound
 log = logging.getLogger('WakaTime')
 
 
-# extensions taking priority over lexer
-EXTENSIONS = {
-    'j2': 'HTML',
-    'markdown': 'Markdown',
-    'md': 'Markdown',
-    'mdown': 'Markdown',
-    'twig': 'Twig',
-}
-
-# lexers to human readable languages
-TRANSLATIONS = {
-    'CSS+Genshi Text': 'CSS',
-    'CSS+Lasso': 'CSS',
-    'HTML+Django/Jinja': 'HTML',
-    'HTML+Lasso': 'HTML',
-    'JavaScript+Genshi Text': 'JavaScript',
-    'JavaScript+Lasso': 'JavaScript',
-    'Perl6': 'Perl',
-    'RHTML': 'HTML',
-}
-
-# extensions for when no lexer is found
-AUXILIARY_EXTENSIONS = {
-    'vb': 'VB.net',
-}
-
-
 def guess_language(file_name):
     """Guess lexer and language for a file.
 
     Returns (language, lexer) tuple where language is a unicode string.
     """
 
+    language = get_language_from_extension(file_name)
+    if language:
+        return language, None
+
     lexer = smart_guess_lexer(file_name)
-
-    language = None
-
-    # guess language from file extension
-    if file_name:
-        language = get_language_from_extension(file_name, EXTENSIONS)
-
-    # get language from lexer if we didn't have a hard-coded extension rule
-    if language is None and lexer:
+    if lexer:
         language = u(lexer.name)
-
-    if language is None:
-        language = get_language_from_extension(file_name, AUXILIARY_EXTENSIONS)
-
-    if language is not None:
-        language = translate_language(language)
 
     return language, lexer
 
@@ -93,14 +56,14 @@ def smart_guess_lexer(file_name):
 
     text = get_file_contents(file_name)
 
-    lexer_1, accuracy_1 = guess_lexer_using_filename(file_name, text)
-    lexer_2, accuracy_2 = guess_lexer_using_modeline(text)
+    lexer1, accuracy1 = guess_lexer_using_filename(file_name, text)
+    lexer2, accuracy2 = guess_lexer_using_modeline(text)
 
-    if lexer_1:
-        lexer = lexer_1
-    if (lexer_2 and accuracy_2 and
-        (not accuracy_1 or accuracy_2 > accuracy_1)):
-        lexer = lexer_2
+    if lexer1:
+        lexer = lexer1
+    if (lexer2 and accuracy2 and
+        (not accuracy1 or accuracy2 > accuracy1)):
+        lexer = lexer2  # pragma: nocover
 
     return lexer
 
@@ -121,7 +84,7 @@ def guess_lexer_using_filename(file_name, text):
     if lexer is not None:
         try:
             accuracy = lexer.analyse_text(text)
-        except:
+        except:  # pragma: nocover
             pass
 
     return lexer, accuracy
@@ -135,44 +98,48 @@ def guess_lexer_using_modeline(text):
 
     lexer, accuracy = None, None
 
-    file_type = get_filetype_from_buffer(text)
+    file_type = None
+    try:
+        file_type = get_filetype_from_buffer(text)
+    except:  # pragma: nocover
+        pass
+
     if file_type is not None:
         try:
             lexer = get_lexer_by_name(file_type)
-        except ClassNotFound:
+        except ClassNotFound:  # pragma: nocover
             pass
 
     if lexer is not None:
         try:
             accuracy = lexer.analyse_text(text)
-        except:
+        except:  # pragma: nocover
             pass
 
     return lexer, accuracy
 
 
-def get_language_from_extension(file_name, extension_map):
-    """Returns a matching language for the given file_name using extension_map.
+def get_language_from_extension(file_name):
+    """Returns a matching language for the given file extension.
     """
 
-    extension = file_name.rsplit('.', 1)[-1] if len(file_name.rsplit('.', 1)) > 1 else None
+    filepart, extension = os.path.splitext(file_name)
 
-    if extension:
-        if extension in extension_map:
-            return extension_map[extension]
-        if extension.lower() in extension_map:
-            return extension_map[extension.lower()]
+    if os.path.exists(u('{0}{1}').format(u(filepart), u('.c'))) or os.path.exists(u('{0}{1}').format(u(filepart), u('.C'))):
+        return 'C'
+
+    extension = extension.lower()
+    if extension == '.h':
+        directory = os.path.dirname(file_name)
+        available_files = os.listdir(directory)
+        available_extensions = list(zip(*map(os.path.splitext, available_files)))[1]
+        available_extensions = [ext.lower() for ext in available_extensions]
+        if '.cpp' in available_extensions:
+            return 'C++'
+        if '.c' in available_extensions:
+            return 'C'
 
     return None
-
-
-def translate_language(language):
-    """Turns Pygments lexer class name string into human-readable language.
-    """
-
-    if language in TRANSLATIONS:
-        language = TRANSLATIONS[language]
-    return language
 
 
 def number_lines_in_file(file_name):
@@ -181,13 +148,18 @@ def number_lines_in_file(file_name):
         with open(file_name, 'r', encoding='utf-8') as fh:
             for line in fh:
                 lines += 1
-    except:
-        return None
+    except:  # pragma: nocover
+        try:
+            with open(file_name, 'r', encoding=sys.getfilesystemencoding()) as fh:
+                for line in fh:
+                    lines += 1
+        except:
+            return None
     return lines
 
 
-def get_file_stats(file_name, notfile=False, lineno=None, cursorpos=None):
-    if notfile:
+def get_file_stats(file_name, entity_type='file', lineno=None, cursorpos=None):
+    if entity_type != 'file':
         stats = {
             'language': None,
             'dependencies': [],
@@ -217,6 +189,10 @@ def get_file_contents(file_name):
     try:
         with open(file_name, 'r', encoding='utf-8') as fh:
             text = fh.read(512000)
-    except:
-        pass
+    except:  # pragma: nocover
+        try:
+            with open(file_name, 'r', encoding=sys.getfilesystemencoding()) as fh:
+                text = fh.read(512000)
+        except:
+            log.exception("Exception:")
     return text