+++ Editor/WakaTime/client/HISTORY.rst
@@ -3,6 +3,62 @@ History
 -------
 
 
+4.1.7 (2015-09-16)
+++++++++++++++++++
+
+- default request timeout of 30 seconds
+- new --timeout command line argument to change request timeout in seconds
+
+
+4.1.6 (2015-09-06)
+++++++++++++++++++
+
+- allow passing command line arguments using sys.argv
+
+
+4.1.5 (2015-09-06)
+++++++++++++++++++
+
+- fix entry point for pypi distribution
+
+
+4.1.4 (2015-08-29)
+++++++++++++++++++
+
+- new --entity and --entitytype command line arguments
+
+
+4.1.3 (2015-08-28)
+++++++++++++++++++
+
+- fix local session caching
+
+
+4.1.2 (2015-08-25)
+++++++++++++++++++
+
+- fix bug in offline caching which prevented heartbeats from being cleaned up
+
+
+4.1.1 (2015-08-25)
+++++++++++++++++++
+
+- send hostname in X-Machine-Name header
+- catch exceptions from pygments.modeline.get_filetype_from_buffer
+- upgrade requests package to v2.7.0
+- handle non-ASCII characters in import path on Windows, won't fix for Python2
+- upgrade argparse to v1.3.0
+- move language translations to api server
+- move extension rules to api server
+- detect correct header file language based on presence of .cpp or .c files named the same as the .h file
+
+
+4.1.0 (2015-06-29)
+++++++++++++++++++
+
+- correct priority for project detection
+
+
 4.0.16 (2015-06-23)
 ++++++++++++++++++
 
@@ -413,4 +469,3 @@ History
 ++++++++++++++++++
 
 - Birth
-